Mapping Oracle TIMESTAMP to java.sql.Timestamp
When Oracle fixed broken type mappings in new (11+) versions of their
database drivers GeoTools had to explicitly map DATE to java.sql.Date
to maintain backwards compatibility with existing GeoTools code. 

Unfortunately, this led to a state when we are unable to read and write
timestamps (i.e. date and time in a single field) using datastores 
based on Oracle database.

This commit introduces a type override that maps java.sql.Timestamp to
Oracle's TIMESTAMP type as discussed on GeoTools Users mailing list:
https://sourceforge.net/p/geotools/mailman/message/35479331/